restructure a bit more
