+++ drive-contents/boot_options/splash_screen.py
