+++ drive-contents/setup/watchshade.py
@@ -38,7 +38,7 @@ class shade(Pages):
             self.page = self.main_shade
 
     def _hide_(self):
-        print(self, self.page)
+        # print(self, self.page)
         if self.page is self.torch_panel:
             self.page = self.main_shade
         super()._hide_()
@@ -54,7 +54,6 @@ class shade(Pages):
             press=lambda: None,
         )
 
-        print(f"Layout._decalrable_={Layout._decalrable_}")
         open_torch = Button(
             text="torch",
             radius=ratio(height // 2),
